Fixing build errors

+++ src/GeometryEditor.cpp
@@ -1,28 +1,19 @@
 ﻿#include "hexa_engine/GeometryEditor.h"
 
-#include "base_lib/Math.h"
+#include <base_lib/Math.h>
 
-Vector3 GeometryEditor::compute_normal(const Vector3& a, const Vector3& b, const Vector3& c)
-{
+Vector3 GeometryEditor::compute_normal(const Vector3& a, const Vector3& b, const Vector3& c) {
     return (c - a).cross_product(b - a).normalized();
 }
 
-void GeometryEditor::optimize(List<StaticMesh::Vertex>& vertices, List<uint>& indices)
-{
-    for (uint i = 0; i < vertices.length() - 1; i++)
-    {
-        for (uint j = i + 1; j < vertices.length(); j++)
-        {
-            if (vertices[i].pos == vertices[j].pos && vertices[i].uv == vertices[j].uv && vertices[i].norm == vertices[j].norm)
-            {
-                for (uint k = 0; k < indices.length(); k++)
-                {
-                    if (indices[k] == j)
-                    {
+void GeometryEditor::optimize(List<StaticMesh::Vertex>& vertices, List<uint>& indices) {
+    for (uint i = 0; i < vertices.length() - 1; i++) {
+        for (uint j = i + 1; j < vertices.length(); j++) {
+            if (vertices[i].pos == vertices[j].pos && vertices[i].uv == vertices[j].uv && vertices[i].norm == vertices[j].norm) {
+                for (uint k = 0; k < indices.length(); k++) {
+                    if (indices[k] == j) {
                         indices[k] = i;
-                    }
-                    else if (indices[k] > j)
-                    {
+                    } else if (indices[k] > j) {
                         indices[k]--;
                     }
                 }
@@ -33,22 +24,14 @@ void GeometryEditor::optimize(List<StaticMesh::Vertex>& vertices, List<uint>& in
     }
 }
 
-void GeometryEditor::optimize_collision(List<StaticMesh::Vertex>& vertices, List<uint>& indices)
-{
-    for (uint i = 0; i < vertices.length() - 1; i++)
-    {
-        for (uint j = i + 1; j < vertices.length(); j++)
-        {
-            if (vertices[i].pos == vertices[j].pos)
-            {
-                for (uint k = 0; k < indices.length(); k++)
-                {
-                    if (indices[k] == j)
-                    {
+void GeometryEditor::optimize_collision(List<StaticMesh::Vertex>& vertices, List<uint>& indices) {
+    for (uint i = 0; i < vertices.length() - 1; i++) {
+        for (uint j = i + 1; j < vertices.length(); j++) {
+            if (vertices[i].pos == vertices[j].pos) {
+                for (uint k = 0; k < indices.length(); k++) {
+                    if (indices[k] == j) {
                         indices[k] = i;
-                    }
-                    else if (indices[k] > j)
-                    {
+                    } else if (indices[k] > j) {
                         indices[k]--;
                     }
                 }
@@ -59,8 +42,7 @@ void GeometryEditor::optimize_collision(List<StaticMesh::Vertex>& vertices, List
     }
 }
 
-void GeometryEditor::remove_indices(List<StaticMesh::Vertex>& vertices, List<uint>& indices)
-{
+void GeometryEditor::remove_indices(List<StaticMesh::Vertex>& vertices, List<uint>& indices) {
     /*for (uint i = 0; i < indices.length() - 1; i++)
     {
         if (indices[i] != i)
@@ -75,8 +57,7 @@ void GeometryEditor::remove_indices(List<StaticMesh::Vertex>& vertices, List<uin
     }*/
 
     List<StaticMesh::Vertex> result;
-    for (auto index : indices)
-    {
+    for (auto index : indices) {
         result.add(vertices[index]);
     }
 
@@ -84,10 +65,8 @@ void GeometryEditor::remove_indices(List<StaticMesh::Vertex>& vertices, List<uin
     indices = List<uint>(indices.length(), [](uint i) -> uint { return i; });
 }
 
-void GeometryEditor::invert_vertices(List<StaticMesh::Vertex>& vertices)
-{
-    for (uint i = 0; i < vertices.length(); i += 3)
-    {
+void GeometryEditor::invert_vertices(List<StaticMesh::Vertex>& vertices) {
+    for (uint i = 0; i < vertices.length(); i += 3) {
         std::swap(vertices[i], vertices[i + 2]);
         vertices[i + 0].norm *= -1;
         vertices[i + 1].norm *= -1;
@@ -95,91 +74,70 @@ void GeometryEditor::invert_vertices(List<StaticMesh::Vertex>& vertices)
     }
 }
 
-void GeometryEditor::invert_indices(List<uint>& indices)
-{
-    for (uint i = 0; i < indices.length(); i += 3)
-    {
+void GeometryEditor::invert_indices(List<uint>& indices) {
+    for (uint i = 0; i < indices.length(); i += 3) {
         std::swap(indices[i], indices[i + 2]);
     }
 }
 
-void GeometryEditor::mirror_x(List<StaticMesh::Vertex>& vertices)
-{
-    for (auto& vertex : vertices)
-    {
+void GeometryEditor::mirror_x(List<StaticMesh::Vertex>& vertices) {
+    for (auto& vertex : vertices) {
         vertex.pos.x *= -1;
         vertex.norm.x *= -1;
     }
 }
 
-void GeometryEditor::mirror_y(List<StaticMesh::Vertex>& vertices)
-{
-    for (auto& vertex : vertices)
-    {
+void GeometryEditor::mirror_y(List<StaticMesh::Vertex>& vertices) {
+    for (auto& vertex : vertices) {
         vertex.pos.y *= -1;
         vertex.norm.y *= -1;
     }
 }
 
-void GeometryEditor::mirror_z(List<StaticMesh::Vertex>& vertices)
-{
-    for (auto& vertex : vertices)
-    {
+void GeometryEditor::mirror_z(List<StaticMesh::Vertex>& vertices) {
+    for (auto& vertex : vertices) {
         vertex.pos.z *= -1;
         vertex.norm.z *= -1;
     }
 }
 
-void GeometryEditor::translate(List<StaticMesh::Vertex>& vertices, const Vector3& offset)
-{
-    for (auto& vertex : vertices)
-    {
+void GeometryEditor::translate(List<StaticMesh::Vertex>& vertices, const Vector3& offset) {
+    for (auto& vertex : vertices) {
         vertex.pos += offset;
     }
 }
 
-void GeometryEditor::rotate(List<StaticMesh::Vertex>& vertices, const Quaternion& quat)
-{
-    for (auto& vertex : vertices)
-    {
+void GeometryEditor::rotate(List<StaticMesh::Vertex>& vertices, const Quaternion& quat) {
+    for (auto& vertex : vertices) {
         vertex.pos = quat.rotate_vector3(vertex.pos);
         vertex.norm = quat.rotate_vector3(vertex.norm);
     }
 }
 
-void GeometryEditor::scale(List<StaticMesh::Vertex>& vertices, const Vector3& factor)
-{
-    for (auto& vertex : vertices)
-    {
+void GeometryEditor::scale(List<StaticMesh::Vertex>& vertices, const Vector3& factor) {
+    for (auto& vertex : vertices) {
         vertex.pos *= factor;
     }
 }
 
-void GeometryEditor::move_to_center(List<StaticMesh::Vertex>& vertices)
-{
+void GeometryEditor::move_to_center(List<StaticMesh::Vertex>& vertices) {
     Vector3 sum;
-    for (const auto& vertex : vertices)
-    {
+    for (const auto& vertex : vertices) {
         sum += vertex.pos;
     }
     sum /= static_cast<float>(vertices.length());
     translate(vertices, sum);
 }
 
-struct Triangle
-{
+struct Triangle {
     uint points[3];
     Vector3 normal;
 
-    bool is_near(const Triangle& triangle)
-    {
+    bool is_near(const Triangle& triangle) {
         byte counter = 0;
-        for (byte i = 0; i < 3; i++)
-        {
-            for (byte j = 0; j < 3; j++)
-            {
-                if (points[i] == triangle.points[j])
-                {
+        for (byte i = 0; i < 3; i++) {
+            for (byte j = 0; j < 3; j++) {
+                if (points[i] == triangle.points[j]) {
                     ++counter;
                     break;
                 }
@@ -189,8 +147,7 @@ struct Triangle
     }
 };
 
-FORCEINLINE float angle(const Vector3& point, const Vector3& front, const Vector3& left)
-{
+FORCEINLINE float angle(const Vector3& point, const Vector3& front, const Vector3& left) {
     float angle = Math::acos_deg(Math::clamp(point.dot_product(front), -1.0f, 1.0f));
     float dot = Math::clamp(point.dot_product(left), -1.0f, 1.0f);
     if (dot > 0)
@@ -198,8 +155,7 @@ FORCEINLINE float angle(const Vector3& point, const Vector3& front, const Vector
     return Math::unwind_angle(angle);
 }
 
-void GeometryEditor::compute_faces(const List<StaticMesh::Vertex>& vertices, const List<uint>& indices, List<Face>& out_faces, List<uint>& out_indices)
-{
+void GeometryEditor::compute_faces(const List<StaticMesh::Vertex>& vertices, const List<uint>& indices, List<Face>& out_faces, List<uint>& out_indices) {
     out_faces.clear();
     out_indices.clear();
 
@@ -207,8 +163,7 @@ void GeometryEditor::compute_faces(const List<StaticMesh::Vertex>& vertices, con
     List<Triangle> triangles(indices.length() / 3);
 
     // calculate normal for each triangle
-    for (uint i = 0; i < indices.length() / 3; i++)
-    {
+    for (uint i = 0; i < indices.length() / 3; i++) {
         triangles[i].points[0] = indices[i * 3 + 0];
         triangles[i].points[1] = indices[i * 3 + 1];
         triangles[i].points[2] = indices[i * 3 + 2];
@@ -216,15 +171,11 @@ void GeometryEditor::compute_faces(const List<StaticMesh::Vertex>& vertices, con
     }
 
     // group triangles by normals
-    for (uint i = 0; i < triangles.length(); i++)
-    {
+    for (uint i = 0; i < triangles.length(); i++) {
         bool face_found = false;
-        for (auto& face : faces)
-        {
-            for (auto face_triangle_index : face)
-            {
-                if (triangles[i].normal == triangles[face_triangle_index].normal && triangles[i].is_near(triangles[face_triangle_index]))
-                {
+        for (auto& face : faces) {
+            for (auto face_triangle_index : face) {
+                if (triangles[i].normal == triangles[face_triangle_index].normal && triangles[i].is_near(triangles[face_triangle_index])) {
                     face_found = true;
                     face.add(i);
                     break;
@@ -235,18 +186,15 @@ void GeometryEditor::compute_faces(const List<StaticMesh::Vertex>& vertices, con
                 break;
         }
 
-        if (!face_found)
-        {
+        if (!face_found) {
             faces.add({i});
         }
     }
 
-    for (auto& face : faces)
-    {
+    for (auto& face : faces) {
         // gather indices for face
         List<uint> face_vertex_indices;
-        for (auto triangle_id : face)
-        {
+        for (auto triangle_id : face) {
             face_vertex_indices.add_unique(triangles[triangle_id].points[0]);
             face_vertex_indices.add_unique(triangles[triangle_id].points[1]);
             face_vertex_indices.add_unique(triangles[triangle_id].points[2]);
@@ -254,17 +202,15 @@ void GeometryEditor::compute_faces(const List<StaticMesh::Vertex>& vertices, con
 
         // calculate face center
         Vector3 center;
-        for (auto face_index : face_vertex_indices)
-        {
+        for (auto face_index : face_vertex_indices) {
             center += vertices[face_index].pos;
         }
-        center /= face_vertex_indices.length();
+        center /= static_cast<float>(face_vertex_indices.length());
         // order indices in counter-clockwise order
         Vector3 orient_axis_f = (vertices[face_vertex_indices[0]].pos - center).normalized();
         Vector3 orient_axis_l = orient_axis_f.cross_product(triangles[face[0]].normal);
         Map<uint, float> angles;
-        for (uint i = 0; i < face_vertex_indices.length(); i++)
-        {
+        for (uint i = 0; i < face_vertex_indices.length(); i++) {
             angles[face_vertex_indices[i]] = angle((vertices[face_vertex_indices[i]].pos - center).normalized(), orient_axis_f, orient_axis_l);
         }
         face_vertex_indices.sort_predicate([&](uint a, uint b) -> bool {
@@ -276,12 +222,10 @@ void GeometryEditor::compute_faces(const List<StaticMesh::Vertex>& vertices, con
     }
 }
 
-void GeometryEditor::compute_normals(const List<StaticMesh::Vertex>& vertices, const List<uint>& indices, List<Vector3>& out_normals, bool invert)
-{
+void GeometryEditor::compute_normals(const List<StaticMesh::Vertex>& vertices, const List<uint>& indices, List<Vector3>& out_normals, bool invert) {
     out_normals.clear();
 
-    for (uint i = 0; i < indices.length() / 3; i++)
-    {
+    for (uint i = 0; i < indices.length() / 3; i++) {
         auto normal = compute_normal(vertices[indices[i * 3 + 0]].pos, vertices[indices[i * 3 + 1]].pos, vertices[indices[i * 3 + 2]].pos);
         if (invert)
             normal *= -1;
@@ -289,12 +233,10 @@ void GeometryEditor::compute_normals(const List<StaticMesh::Vertex>& vertices, c
     }
 }
 
-void GeometryEditor::compute_normals(List<StaticMesh::Vertex>& vertices, const List<uint>& indices, bool invert)
-{
+void GeometryEditor::compute_normals(List<StaticMesh::Vertex>& vertices, const List<uint>& indices, bool invert) {
     List<List<Vector3>> vertex_normals = List<List<Vector3>>(vertices.length());
 
-    for (uint i = 0; i < indices.length() / 3; i++)
-    {
+    for (uint i = 0; i < indices.length() / 3; i++) {
         const uint i0 = indices[i * 3 + 0];
         const uint i1 = indices[i * 3 + 1];
         const uint i2 = indices[i * 3 + 2];
@@ -306,39 +248,31 @@ void GeometryEditor::compute_normals(List<StaticMesh::Vertex>& vertices, const L
         vertex_normals[i2].add(normal);
     }
 
-    for (uint i = 0; i < indices.length(); i++)
-    {
+    for (uint i = 0; i < indices.length(); i++) {
         const uint index = indices[i];
 
         Vector3 normal_sum;
-        for (auto norm : vertex_normals[index])
-        {
+        for (auto norm : vertex_normals[index]) {
             normal_sum += norm;
         }
-        vertices[index].norm = (normal_sum / vertex_normals.length()).normalized();
+        vertices[index].norm = (normal_sum / static_cast<float>(vertex_normals.length())).normalized();
     }
 }
 
-void GeometryEditor::remove_unused_vertices(List<StaticMesh::Vertex>& vertices, List<uint>& indices)
-{
+void GeometryEditor::remove_unused_vertices(List<StaticMesh::Vertex>& vertices, List<uint>& indices) {
     List<bool> referred = List<bool>(vertices.length(), false);
 
-    for (auto index : indices)
-    {
+    for (auto index : indices) {
         referred[index] = true;
     }
 
-    for (uint i = 0; i < vertices.length(); i++)
-    {
-        if (!referred[i])
-        {
+    for (uint i = 0; i < vertices.length(); i++) {
+        if (!referred[i]) {
             vertices.remove_at(i);
             referred.remove_at(i);
 
-            for (uint j = 0; j < indices.length(); j++)
-            {
-                if (indices[j] > i)
-                {
+            for (uint j = 0; j < indices.length(); j++) {
+                if (indices[j] > i) {
                     indices[j--]--;
                 }
             }
@@ -346,11 +280,9 @@ void GeometryEditor::remove_unused_vertices(List<StaticMesh::Vertex>& vertices, 
     }
 }
 
-Shared<StaticMesh> GeometryEditor::get_unit_cube()
-{
+Shared<StaticMesh> GeometryEditor::get_unit_cube() {
     static Shared<StaticMesh> result;
-    if (result == nullptr)
-    {
+    if (result == nullptr) {
         List<StaticMesh::Vertex> vertices = {
             // down
             {{0, 0, 0}, {0, 1}, -Vector3::up()},
